requirements.txt
python-dotenv
google-genai
transformers
torch
datasets
torchaudio
Pillow